objectmodel/db/NodeManager: make helma work with HSSQLDB and SQLITE
HSSQLDB and SQLITE use uppercase letters for field names.
Previous to this commit, helma compared the field names with lower case
letters to find the prototype field. This failed for databases
with names in uppercase.

Now comparison is performed case insensitive and hence the prototype
field can be found on any database!